Implemented Embedded Google Maps on the campground show page.
installed geocoder to assist with finding geocoordinates. The embedded map shows up on the left pane of the show page. Other small edits also took place in styling the New and Edit Pages for the campgrounds to allow for additional fields needed in google maps/geocoder and then of course also had to change the data model for campgrounds to allow this.